dev -> staging (#144)
* feature: adds position designations to cycle position stats. TM-695

* fix: filter designations by cycle and position

* fix: 404 if bidcycle or position cannot be found

* feature: handle errors that fsbid may raise when submitting a bid

* adds endpoint to view list of logs available and get the contents of a particular log

* remove unnecessary imports

* log entry model managed should be false

* fixing linting issues

* additional lint errors fixed

* adding endpoints for running data sync ad hoc

* forcing superuser permission and removing unused imports and functions

* fix: remove posted date from PV. TM-909 (#94)

* feature: add endpoints for getting and updating the about page content. TM-912 (#93)

* feature: add endpoints for getting and updating the about page content. TM-912

* chore: update comment

* Fix/realign pv values. TM-909 (#96)

* fix: remove posted date from PV. TM-909

* fix: projected vacancy data matching new FSBid endpoints. TM-909

* fix: a few updates from pr feedback

* added endpoint to run update_relationships command (#95)

* adding endpoint for update_string_representations command (#98)

* adding endpoint for update_string_representations command

* lint issue

* feature: allow for saving projected vacancies searches. TM-642

* fix: properly parse langualge string. correct field name error

* fix: remove null entries from lang. list

* fix: syntax

* feature: pv filtering for bureau hierarchy. TM-950

* chore: missed files

* chore: missed this as well

* Update apps.py

* fix: strip the end of the language text so there is no extra spaces

* chore: fix test data

* fix: fix PV favorites by using the id of the pv

* fix: send response object when no results are returned

* adding endpoints to edit SynchronizationJobs in DB and reset SynchronizationJob

* add use_last_date_updated as writable field, remove unnecesary filters

* removing reference to deleted filter class

* fix: correctly filter by location codes. TM-949

* fix: update the shape of the pv object to match new position shape (#109)

* fix: update availability to use cycle position to allow bidding

* updating response from data_sync list for conistency

* fix: use correct filter synatx

* chore: add cycle positions to the demo env setup

* chore: properly ref. the user first_name

* fix: process multiple filter values

* fix: position designations on cycle positions

* chore: fix formatting on files that are unchanged

* fix: query bids correctly for determining availability

* Create SynchronizationTask (#29)

* Create SynchronizationTask

* linting

* Testing saving task

* Fix merge issue

* Delete migration

* Re-make migration

* Remove Flask (#123)

* Linting (#115)

* Linting

* Linting

* removing historical assignments (#124)

* chore: update the demo environment task. TM-1115

* chore: add TED to cycle position

* save last_synchronization at beginning of process instead of end (#128)

* Added string representation to CyclePosition (#127)

* adding JWT to header for all calls to FSBid (#129)

* adding JWT to header for all calls to FSBid

* updating tests to pretend to use JWT

* Update/ad id in fsbid calls (#131)

* added ad_id to all fsbid calls

* escaping characters properly and updating url

* 🐛 Fix merge error 🐛

* 🐛 Fix merge 🐛

* adding admin view to check status of syncs

* cleaning out the dryer lint

* removing unnecessary pk value

* ESSR fixes for dev testing (#135)

* ESSR fixes for dev testing

* adding nosec

* fixing lint issues

* additional placeholder JWT

* adding obc_url to country and post (#136)

* adding obc_url to country and post

* Remove trailing slash

* Fix url slug

* chore: merge migrations to resolve conflict

* adding endpoints to track logins and unique logins (#139)

* adding endpoints to track logins and unique logins

* change submit to post, some additional cleanup

* Trigger notification

* ad_id -> fv_request_params.ad_id (#138)

* fix: do not use the bidcycle and position to determine uniquiness on a cycle position (#134)

* Override CORS headers

* fix: rename the jwt header sent to FSBid to match their naming. always send ad_id to the fsbid reqeuests

* Feature/jwtdecode (#143)

* decode jwt

* split services

* split services and update calls

* fix typo

* decode jwt

* split services

* split services and update calls

* fix typo

* chore: cleanup some tests and references to fake jwt that do not decode

* remove services file

* remove ad_id decode and header as it is not needed by fsbid
