+++ talentmap_api/integrations/serializers.py
@@ -3,7 +3,7 @@ from rest_framework.reverse import reverse
 from django.apps import apps
 
 from talentmap_api.common.serializers import PrefetchedSerializer, StaticRepresentationField
-from talentmap_api.integrations.models import SynchronizationJob
+from talentmap_api.integrations.models import SynchronizationJob, SynchronizationTask
 
 
 class SynchronizationJobSerializer(PrefetchedSerializer):
@@ -12,3 +12,10 @@ class SynchronizationJobSerializer(PrefetchedSerializer):
         model = SynchronizationJob
         fields = "__all__"
         writable_fields = ('last_synchronization', 'next_synchronization', 'delta_synchronization', 'running', 'talentmap_model', 'priority', 'use_last_date_updated')
+
+
+class SynchronizationTaskSerializer(PrefetchedSerializer):
+
+    class Meta:
+        model = SynchronizationTask
+        fields = "__all__"
