+++ talentmap_api/integrations/synchronization_helpers.py
@@ -539,7 +539,9 @@ def mode_cycle_positions(last_updated_date=None):
 
         if position:
             updated_instances.append(position)
-            cycle_position, _ = CyclePosition.objects.get_or_create(bidcycle=bc, position=position, _cp_id=data["CP_ID"])
+            cycle_position, _ = CyclePosition.objects.get_or_create(_cp_id=data["CP_ID"])
+            cycle_position.position = position
+            cycle_position.bidcycle = bc
             cycle_position.status_code = data["STATUS_CODE"]
             cycle_position.status = data["STATUS"]
             cycle_position.created = ensure_date(data["DATE_CREATED"], utc_offset=-5)
