+++ talentmap_api/integrations/urls/data_sync_actions.py
@@ -13,6 +13,7 @@ urlpatterns = [
     url(r'^reset/(?P<pk>[0-9]+)/$', views.DataSyncScheduleResetView.as_view(), name="integrations.DataSyncScheduleActionView"),
     url(r'^updatestringrepresentations/$', views.UpdateStringRepresentationsActionView.as_view(), name="integrations.UpdateStringRepresentationsActionView"),
     url(r'^updaterelationships/$', views.UpdateRelationshipsActionView.as_view(), name="integrations.UpdateRelationshipsActionView"),
+    url(r'^status/$', views.DataSyncTaskListView.as_view(get_list), name="integrations.SynchronizationTask-list-status"),
     url(r'', views.DataSyncListView.as_view(), name="integrations.DataSyncListView"),
 ]
