+++ talentmap_api/stats/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class StatsConfig(AppConfig):
+    name = 'stats'
